Pull out operation processing into its own class
This lets us set context at the class level per operation (like identity
headers).
